Patched exception via ...::client::connect()
Calling disconnect without checking whether the client is actually connected brings the (highly probable) risk of the disconnect function throwing an exception given how the client typically only calls disconnect when there's a chance that it isn't connected at all.
This is fixed by adding a simple check, identical to the one in T2::net::client::disconnect, to prevent an exception being thrown.